Revert "Revert "move modules and classes from lib to models that requires autoload""
This reverts commit f348bca361cf960ab742c4f3597c86fb1e4b2770.
