+++ app/models/concerns/yeti/network_detector.rb
