+++ app/models/concerns/yeti/resource_status.rb
@@ -1,0 +1,28 @@
+module Yeti
+  module ResourceStatus
+    extend ActiveSupport::Concern
+    included do
+      scope :disabled, -> { base.where enabled: false }
+      scope :enabled, -> { base.where enabled: true }
+
+      def status_sym
+        self.enabled? ? :enabled : :disabled
+      end
+
+      def enable!
+        self.enabled = true
+        save!
+      end
+
+      def disable!
+        self.enabled = false
+        save!
+      end
+
+      def disabled?
+        !self.enabled
+      end
+
+    end
+  end
+end
