+++ app/models/invoice_generator.rb
