+++ app/models/yeti/cdrs_filter.rb
