+++ app/models/yeti/outgoing_registrations.rb
